chore: (processm.core.verifiers.causalnet.CausalNetVerifierImpl) some further optimization
Each partial sequence now has a map for its states: from the hash code
of the unique set of a state to the collection of the states. This
greately facilitates looking for boring subsets, because we compare only
with these states that have the unique set's hash code the same as the
new state's.