chore: (processm.conformance.models.alignments) compare events with == instead of ===
The purpose of using === instead of == was to distinguish different
occurrences of events during the alignment, but with Event.equals() working
correctly this is not necessary. Confusing two events with each other
while merging alignments created within DecompositionAligner seems to be
impossible, as a decomposition always contain all events with the same
name and their order is maintained.


Former-commit-id: 6111817ecbd92128285d8f9c02c26dc5be2ba8fb