fix: PSQLException: ERROR: index row size 2792 exceeds btree version 4 maximum 2704 for index "_hyper_7_2_chunk_logs_attributes_key" (closes #101)
Replaced the BTree indexes on the (logs/traces/events)_attributes.key columns with the hash indexes. This should have also a positive impact on the performance of the PQL interpreter.
The above fix still does not allow storing the Hospital_Log.xes.gz, as the number of parameters in an SQL query exceeds the PostgreSQL limit (the same error as in #102).


Former-commit-id: d21da24dbd1d9f416029807f3ff976d62cb453c4