feat: add AppendingDBXESOutputStream (#110)
AppendingDBXESOutputStream writes the supplied XES components at the ends of the corresponding components existing in the database if any. It appends an event to the end of its parent trace and the trace to the end of its parent log. If the parent event/log does not exist is will be created. AppendingDBXESOutputStream uses the "identity:id" attribute to uniquely identify the XES components.


Former-commit-id: d941b8af968765c1bbb40740c9c7fc013247d74b