fix: InferConceptInstanceFromStandardLifecycle loses some attributes in the modified events
The loss is caused by the property initializers of the inherited properties from XESComponent in EventOrTraceBase that are called after the constructor of XESComponent. The constructor of XESComponent fills the standard XES properties from the map given as the argument, and the initializers then override those values. The workaround relies on calling Event.setStandardAttributes() with the name mapping read from log extensions.


Former-commit-id: 70d261e05ed200c1b74081f6d42a3d49b178421c