fix: (processm.services.api.WorkspacesApiTest) fixing #170 and #182
I was able to reproduce problems from TeamCity on a local system by
limiting the number of available CPU cores with schedtool, e.g.,
`schedtool -a 0x3 -e mvn -Dtest='processm.services.api.WorkspacesApiTest'`
test The proposed patch seems to work correctly with schedtool.
According to the documentation [1], the default dispatcher "It is backed
by a shared pool of threads on JVM and Native. By default, the maximum
number of threads used by this dispatcher is equal to the number
of CPU cores, but is at least two." I think that some code that looks
like it should suspend the coroutine and release the thread back to the
pool, does not do it (fast enough?), in turn leading to deadlock-like
behaviours. IO dispatcher, on the other hand, has by default at least 64
threads, which should suffice I hope.

[1] https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-dispatchers/-default.html


Former-commit-id: c65c5835bb749fc3b62460b4f5e0e585ba653713