chore: (processm.etl) replacing AutomaticEtlProcessRelations with a reference to Relationships (#175)
AutomaticEtlProcessRelations was a table redundant to Relationships, but
without all the necessary informations. While it could be argued that it
offered greater flexibility and it simply should have been expanded to
include missing information, I saw no point in it, as the current
structure relies on the metamodel anyhow (table Classes).


Former-commit-id: 3a7819383a257fc64fa0dad2efbbfb30529d2470