chore: (processm.services) Splitting CaseNotion in the API spec into two types: one to represent relationship graph, and the other the represent case notions (#175)
Previously, the single type had many optional fields and one had to
ensure they are filled when necessary etc. At the same time, the same
data was transferred multiple times, even though it was unnecessary.
In addition to that, some unnecessary rewriting from one data structure
to another in the backend was removed.


Former-commit-id: b739f8272ed2c940a0747a40143c339d4ce74de0