fix: build and some tests crash on macos with Docker Deskop 4.33.0 or newer
The source of the issue is a known bugin mssql-server-docker that causes mssql-server fail with the symptoms you are describing when running on Linux kernels 6.7 and later.
Docker Desktop 4.33 ships with Linux Kernel 6.10, while Docker Desktop 4.32 is on kernel 6.6.

Microsoft has fixed the bug on mcr.microsoft.com/mssql/server starting from tag 2022-CU11-ubuntu-22.04 (and 2022-latest), but the fix appears not to have been backported to 2017-* tags.
https://github.com/docker/for-mac/issues/7368

Former-commit-id: 1fc4da629bc3e8144bf9d2d4f3bdb02cd1042486